+++ apps/chat/src/app/pages/main/index.tsx
@@ -1,9 +1,7 @@
-import { ModalCreateClan, ModalListClans, NavLinkComponent, SearchModal } from '@mezon/components';
+import { ForwardMessageModal, MessageModalImage, ModalCreateClan, ModalListClans, NavLinkComponent, SearchModal } from '@mezon/components';
 import { useApp, useAppNavigation, useAppParams, useFriends, useMenu, useMessageValue, useReference } from '@mezon/core';
-import { selectAllClans, selectCurrentChannel, selectCurrentClan, directActions, useAppDispatch, selectDirectById, selectAllDirectMessages, selectDmGroupCurrentId, selectDirectsUnreadlist, selectCloseMenu, selectStatusMenu } from '@mezon/store';
+import { selectAllClans, selectCurrentChannel, selectCurrentClan, selectDirectsUnreadlist, selectCloseMenu, selectStatusMenu, selectOpenModalAttachment } from '@mezon/store';
 import { Image } from '@mezon/ui';
-import ForwardMessageModal from 'libs/components/src/lib/components/ForwardMessage';
-import MessageModalImage from 'libs/components/src/lib/components/MessageWithUser/MessageModalImage';
 import { getIsShowPopupForward, toggleIsShowPopupForwardFalse } from 'libs/store/src/lib/forwardMessage/forwardMessage.slice';
 import { useCallback, useEffect, useState } from 'react';
 import { useModal } from 'react-modal-hook';
@@ -103,6 +101,7 @@ function MyApp() {
 	const { appearanceTheme } = useApp();
 	const { setMode } = useMessageValue();
 	const { setOpenOptionMessageState } = useReference();
+	const openModalAttachment = useSelector(selectOpenModalAttachment);
 
 	const handleClick = useCallback(() => {
 		setOpenOptionMessageState(false);
@@ -189,7 +188,7 @@ function MyApp() {
 				</div>
 			</div>
 			<MainContent />
-			<MessageModalImage />
+			{openModalAttachment && <MessageModalImage />}
 		</div>
 	);
 }
